requirements.txt
# Core dependencies
opencv-python==4.9.0.80
numpy==1.26.4
PyQt5==5.15.10
PyYAML==6.0.1
python-dotenv==1.0.0
pillow==10.2.0
pyautogui==0.9.54
pynput==1.7.6
psutil==5.9.8
requests==2.31.0
colorama==0.4.6
mss==9.0.1
pygetwindow==0.0.9
pyinstaller==6.11.1